fix some warning
git-svn-id: http://svn.tls.cena.fr/svn/ivy/ivy-c/trunk@3467 b9a2a475-3c1e-4ae0-83c4-c96c2d81b069

+++ tools/ivyperf.c
@@ -44,7 +44,7 @@ const char *mymessages[] = { "IvyPerf", "ping", "pong" };
 static double origin = 0;
 int nbMsgReceive=0;
 int nbMsgEmit=0;
-int nbMsg = 10;
+long nbMsg = 10;
 
 
 double minRoundTrip=1e12;
